8_advance_datapipeline/requirements.txt
google-cloud-pubsub
google-cloud-bigquery